feat(database): change score fields in ScoreReport model from Int to Float data type for more precise calculations and storage feat(schema): update ScoreReportSchema in Zod to reflect the changes in the database schema for consistency and accuracy

+++ web/src/lib/schema/zod/index.ts
@@ -110,9 +110,9 @@ export type CarbonEmission = z.infer<typeof CarbonEmissionSchema>
 export const ScoreReportSchema = z.object({
   id: z.string().uuid(),
   carbonEmissionId: z.string(),
-  scoreCO2Reduction: z.number().int(),
-  scoreCarbonEfficiency: z.number().int(),
-  totalScore: z.number().int(),
+  scoreCO2Reduction: z.number(),
+  scoreCarbonEfficiency: z.number(),
+  totalScore: z.number(),
 })
 
 export type ScoreReport = z.infer<typeof ScoreReportSchema>
